requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
psycopg2-binary
python-jose
passlib==1.7.4
bcrypt==4.0.1
pydantic
pydantic-settings
openai
email-validator
httpx
Pillow
python-multipart